Try speeding up the build some more
The number of tests we run is increasing
