+++ src/test/java/io/netty/buffer/api/BufferBulkAccessTest.java
@@ -303,60 +303,6 @@ public class BufferBulkAccessTest extends BufferTestSupport {
         }
     }
 
-    private static final Memoize<Fixture[]> OTHER_FIXTURES = new Memoize<Fixture[]>(
-            () -> Arrays.stream(allocators()).filter(filterOfTheDay(10)).toArray(Fixture[]::new));
-
-    @ParameterizedTest
-    @MethodSource("allocators")
-    public void writeBytesMustTransferDataAndUpdateOffsets(Fixture fixture) {
-        try (BufferAllocator alloc1 = fixture.createAllocator()) {
-            // Only test 10% of available combinations. Otherwise, this takes too long.
-            Fixture[] allocators = OTHER_FIXTURES.get();
-            Arrays.stream(allocators).parallel().forEach(otherFixture -> {
-                try (BufferAllocator alloc2 = otherFixture.createAllocator();
-                     Buffer target = alloc1.allocate(37);
-                     Buffer source = alloc2.allocate(35)) {
-                    // BE to BE
-                    target.order(BIG_ENDIAN);
-                    source.order(BIG_ENDIAN);
-                    verifyWriteBytes(target, source);
-
-                    // LE to BE
-                    target.fill((byte) 0).reset().order(BIG_ENDIAN);
-                    source.fill((byte) 0).reset().order(LITTLE_ENDIAN);
-                    verifyWriteBytes(target, source);
-
-                    // BE to LE
-                    target.fill((byte) 0).reset().order(LITTLE_ENDIAN);
-                    source.fill((byte) 0).reset().order(BIG_ENDIAN);
-                    verifyWriteBytes(target, source);
-
-                    // LE to LE
-                    target.fill((byte) 0).reset().order(LITTLE_ENDIAN);
-                    source.fill((byte) 0).reset().order(BIG_ENDIAN);
-                    verifyWriteBytes(target, source);
-                } catch (Exception e) {
-                    e.addSuppressed(new RuntimeException("other fixture was: " + otherFixture));
-                    throw e;
-                }
-            });
-        }
-    }
-
-    private static void verifyWriteBytes(Buffer target, Buffer source) {
-        for (int i = 0; i < 35; i++) {
-            source.writeByte((byte) (i + 1));
-        }
-        target.writeBytes(source);
-        assertThat(target.readerOffset()).isZero();
-        assertThat(target.writerOffset()).isEqualTo(35);
-        assertThat(source.readerOffset()).isEqualTo(35);
-        assertThat(source.writerOffset()).isEqualTo(35);
-        try (Buffer readableSlice = target.slice()) {
-            assertEquals(source, readableSlice);
-        }
-    }
-
     @ParameterizedTest
     @MethodSource("allocators")
     public void writeBytesMustWriteAllBytesFromByteArray(Fixture fixture) {
